# openapi.json
{"openapi":"3.0.0","paths":{"/call-center/auth/login":{"post":{"operationId":"AuthCallCenterController_login","summary":"Login the call center","description":"Return the bearer token for the call center, the email is verified only if the admin was logged in before.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginCallCenterRequest"}}}},"responses":{"200":{"description":"The bearer token for the admin Login","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginCallCenterResponse"}}}}},"tags":["Auth"]}},"/call-center/emergency":{"get":{"operationId":"EmergencyCallCenterController_getEmergency","summary":"Get all emergencies that your previously created","parameters":[],"responses":{"200":{"description":"The emergencies were found.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EmergencyInfoResponse"}}}}}},"tags":["Emergency"],"security":[{"bearer":[]}]},"post":{"operationId":"EmergencyCallCenterController_createEmergency","summary":"Create a new emergency","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateNewEmergencyRequest"}}}},"responses":{"200":{"description":"The emergency was created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmergencyInfoResponse"}}}}},"tags":["Emergency"],"security":[{"bearer":[]}]}}},"info":{"title":"StayAlive API (Call Center)","description":"StayAlive API description for call center","version":"1.0","contact":{"name":"Team StayAlive","url":"https://www.stayalive.fr","email":"stayalive.eip@gmail.com"}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginCallCenterRequest":{"type":"object","properties":{"email":{"type":"string","description":"The email of the call center.","example":"jophn@doe.net"},"password":{"type":"string","description":"The password of the call center.","example":"myPassword123"}},"required":["email","password"]},"LoginCallCenterResponse":{"type":"object","properties":{"token":{"type":"string","description":"The token of the call center.","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Ik5vbmUiLCJpYXQiOjE2MjU2NjMxNjksImV4cCI6MTYyNTY2NjM2OX0.6Ym0rQy8q1o9Yz3XsV4YKk1bDn4N1q4w6d8LgX8qj0A"}},"required":["token"]},"ObjectId":{"type":"object","properties":{}},"EmergencyInfoResponse":{"type":"object","properties":{"id":{"$ref":"#/components/schemas/ObjectId"}},"required":["id"]},"CreateNewEmergencyRequest":{"type":"object","properties":{"info":{"type":"string","description":"The info of the emergency, this information will be shown to the rescuer"},"position":{"type":"string","description":"The position of the emergency, this information will be shown to the rescuer","example":{"lat":"123.123","long":"123.123"}}},"required":["info","position"]}}}}